+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Version history for hwk
 
+## 0.2.1 (2020-10-11)
+- add --line mode: takes a function on a String - applied to every line
+  - uses new ToString class (not ToList)
+- add --type-check: prints the type of a given function
+
 ## 0.2 (2020-10-10)
 - first release by Jens Petersen
 - uses hint library and Hwk configuration module